requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
autopep8==2.0.4
BareNecessities==0.2.8
beautifulsoup4==4.12.2
blinker==1.7.0
bokeh==3.3.1
boto3==1.29.3
botocore==1.32.3
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
contourpy==1.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
DateTime==5.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==4.1.12
django-cors-headers==4.3.0
django-filter==23.3
django-multiselectfield==0.1.12
django-phone-field==1.8.1
django-rest-swagger==2.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
dnspython==2.4.2
drf-yasg==1.21.7
executing==2.0.1
Faker==19.12.0
filelock==3.12.4
Flask==3.0.0
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
h11==0.14.0
httplib2==0.22.0
icalendar==5.0.10
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipython==8.18.1
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jose==1.0.0
JSON-log-formatter==0.5.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
Mail==2.1.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mongoengine==0.27.0
nltk==3.8.1
numpy==1.26.2
oauth2client==4.1.3
oauthlib==3.2.2
openapi==1.1.0
openapi-codec==1.3.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
parso==0.8.3
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
prompt-toolkit==3.0.41
protobuf==4.24.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycodestyle==2.11.1
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pymongo==3.12.3
pyparsing==3.1.1
pypiwin32==223
pyproject-api==1.6.1
PySocks==1.7.1
pytest==7.4.3
pytest-django==4.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-nmap==0.7.1
python3-openid==3.2.0
pyttsx3==2.90
pytube==15.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rpds-py==0.13.1
rsa==4.9
s3transfer==0.7.0
selenium==4.15.1
setuptools==69.0.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.0
sqlparse==0.2.4
stack-data==0.6.3
tornado==6.4
tox==4.11.3
tqdm==4.66.1
traitlets==5.14.0
trio==0.23.0
trio-websocket==0.11.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.24.5
virtualenvwrapper-win==1.2.7
wcwidth==0.2.12
Werkzeug==3.0.1
whatever==0.7
wheel==0.42.0
whitenoise==6.6.0
wsproto==1.2.0
xyzservices==2023.10.1
zope.interface==6.1
